split translate_diann
