+++ docs/spectral_library/translate.rst
@@ -1,9 +1,8 @@
 alphabase.spectral_library.translate
 =====================================
 
-Export a spectral library to tsv (:func:`translate_to_tsv`) or to a DIA-NN 1.9.1+ ``.parquet``
-library (:func:`translate_to_parquet`); see the
-:doc:`DIA-NN parquet notebook <../nbs/diann_parquet_library>`.
+Export a spectral library to tsv (:func:`translate_to_tsv`). For DIA-NN ``.parquet``
+export see :doc:`translate_diann <translate_diann>`.
 
 .. automodule:: alphabase.spectral_library.translate
    :members:
