+++ docs/spectral_library/translate_diann.rst
@@ -1,0 +1,11 @@
+alphabase.spectral_library.translate_diann
+===========================================
+
+Write an alphabase library to a DIA-NN 1.9.1+ ``.parquet`` spectral library
+(:func:`translate_to_parquet`); see the
+:doc:`DIA-NN parquet notebook <../nbs/diann_parquet_library>`.
+
+.. automodule:: alphabase.spectral_library.translate_diann
+   :members:
+   :undoc-members:
+   :show-inheritance:
